Add exception logging.
mtb-daemon allows to log exception to separate files (1 file = 1 exception).
This is suitable mainly for production deployment, where "usual" logs
are ignored, however when exception occurs loglevel is increased and
some messages in history & future are also logged. This allows for
inspection of the cause of the exception. Iy is intended mainly for
diagnostics in production.

See "production_logging" in mtb-daemon.json.